+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [1.0.29-cloud] - 2024-06-07
 ### Changed
 - First part of the sync (conjur cloud with conjur enterprise), support for arm64 + cli version 8 + Add ARM64 packages build
+- Issuer API will not return secret_access_key anymore, instead they will return '*****'
 
 ## [1.0.28-cloud] - 2024-06-02
 ### Changed